+++ test.sh
@@ -7,9 +7,11 @@ bin/falderal -v -t || exit 1
 cd tests
 
 FIRST_TESTS="
-test1 test2 test3 test4 test5 test9
-test-bad-indentation test-input-sections test-utf8 test-crlf
-test-err-no-stderr test-shared-body
+test1 test2 test3 test4 test5
+test-utf8 test-crlf
+test-bad-indentation
+test-input-sections test-shared-body
+test-stdout-stderr test-err-no-stderr
 "
 for TEST in ${FIRST_TESTS}; do
     echo ${TEST}...